# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/handler/{handler_id}/members": {
            "put": {
                "description": "Update Handler Member Role",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Handlers"
                ],
                "summary": "Update Handler Member Role",
                "parameters": [
                    {
                        "description": "HandleOptions",
                        "name": "HandlerMemberEditForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.HandlerMemberEditForm"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "Create Handler Member",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Handlers"
                ],
                "summary": "Create Handler Member",
                "parameters": [
                    {
                        "description": "HandleOptions",
                        "name": "HandlerMemberCreateForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.HandlerMemberCreateForm"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Delete Handler Member",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Handlers"
                ],
                "summary": "Delete Handler Member",
                "parameters": [
                    {
                        "description": "HandleOptions",
                        "name": "HandlerMemberDeleteForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.HandlerMemberDeleteForm"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/handlers": {
            "get": {
                "description": "Get Handlers list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Handlers"
                ],
                "summary": "Handlers",
                "responses": {}
            },
            "post": {
                "description": "Update Handler information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Handlers"
                ],
                "summary": "Create Handler",
                "parameters": [
                    {
                        "description": "HandleOptions",
                        "name": "HandlerCreateForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.HandlerCreateForm"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/handlers/{handler_id}": {
            "get": {
                "description": "Get Handler information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Handlers"
                ],
                "summary": "Handlers",
                "responses": {}
            },
            "put": {
                "description": "Update Handler information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Handlers"
                ],
                "summary": "Handlers",
                "parameters": [
                    {
                        "description": "HandleOptions",
                        "name": "HandlerUpdateForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.HandlerUpdateForm"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Delete Handler",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Handlers"
                ],
                "summary": "Handlers",
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "api.HandlerCreateForm": {
            "type": "object",
            "required": [
                "dom",
                "domain",
                "path",
                "screenshot"
            ],
            "properties": {
                "dom": {
                    "type": "boolean"
                },
                "domain": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "screenshot": {
                    "type": "boolean"
                }
            }
        },
        "api.HandlerMemberCreateForm": {
            "type": "object",
            "required": [
                "email",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.PermisionType"
                }
            }
        },
        "api.HandlerMemberDeleteForm": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "api.HandlerMemberEditForm": {
            "type": "object",
            "required": [
                "role",
                "user_id"
            ],
            "properties": {
                "role": {
                    "$ref": "#/definitions/models.PermisionType"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "api.HandlerUpdateForm": {
            "type": "object",
            "required": [
                "Dom",
                "Screenshot"
            ],
            "properties": {
                "CollectedPages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "Dom": {
                    "type": "boolean"
                },
                "Screenshot": {
                    "type": "boolean"
                }
            }
        },
...
        "models.PermisionType": {
            "type": "string",
            "enum": [
                "READ",
                "WRITE",
                "OWNER"
            ],
            "x-enum-varnames": [
                "READ",
                "WRITE",
                "OWNER"
            ]
        }
    }
}